+++ src/www/html/mod.rs
@@ -10,13 +10,14 @@ mod from_str;
 mod named;
 mod test;
 
-use crate::colour::{Colour, Rgba};
+use crate::{Colour, DefinedGamut};
+use crate::rgb::{SRgba, SRgb};
 
 use core::fmt::{self, Debug, Display, Formatter};
 
-/// A CSS colour.
+/// An HTML colour.
 ///
-/// This type is guaranteed to have the exact same layout and representation as <code>[Rgba]&lt;[u8]&gt;</code> in permanent [`sRGB`](https://en.wikipedia.org/wiki/SRGB), and -- for all intends and purposes -- these types are also equivalent.
+/// This type is guaranteed to have the exact same layout and representation as <code>[SRgba]&lt;[u8]&gt;</code>, and -- for all intends and purposes -- these types are also equivalent.
 ///
 /// Where this type differs is in that it may be more suited in user interactions due to its [`FromStr`](core::str::FromStr) and [`Display`] implementations.
 #[repr(transparent)]
@@ -24,18 +25,18 @@ use core::fmt::{self, Debug, Display, Formatter};
 #[cfg_attr(feature = "bytemuck", derive(bytemuck::Pod, bytemuck::Zeroable))]
 #[cfg_attr(feature = "serde",    derive(serde::Deserialize, serde::Serialize))]
 #[cfg_attr(feature = "zerocopy", derive(zerocopy::FromZeros, zerocopy::Immutable, zerocopy::IntoBytes))]
-pub struct Css(Rgba<u8>);
+pub struct Html(SRgba<u8>);
 
-impl Css {
-	/// Constructs a new CSS colour.
+impl Html {
+	/// Constructs a new HTML colour.
 	#[inline(always)]
 	#[must_use]
 	pub const fn new(red: u8, green: u8, blue: u8, alpha: u8) -> Self {
-		let colour = Rgba::new(red, green, blue, alpha);
-		Self::from_rgba(colour)
+		let colour = SRgba::new(red, green, blue, alpha);
+		Self::from_s_rgba(colour)
 	}
 
-	/// Constructs a new CSS colour from [`u32`].
+	/// Constructs a new HTML colour from [`u32`].
 	///
 	/// The `u32` value is reinterpreted as a four contiguous `u8` objects corresponding to each of the four channels.
 	#[inline(always)]
@@ -45,23 +46,23 @@ impl Css {
 		Self::new(red, green, blue, alpha)
 	}
 
-	/// Converts an SRGBA colour into a CSS colour.
+	/// Converts an SRGBA colour into an HTML colour.
 	///
 	/// This conversion is always lossless and zero-cost.
 	#[inline(always)]
 	#[must_use]
-	pub const fn from_rgba(colour: Rgba<u8>) -> Self {
+	pub const fn from_s_rgba(colour: SRgba<u8>) -> Self {
 		Self(colour)
 	}
 
-	/// Deconstructs a CSS colour.
+	/// Deconstructs an HTML colour.
 	#[inline(always)]
 	#[must_use]
 	pub const fn get(self) -> (u8, u8, u8, u8) {
-		self.to_rgba().get()
+		self.to_s_rgba().get()
 	}
 
-	/// Converts a CSS colour to [`u32`].
+	/// Converts an HTML colour to [`u32`].
 	///
 	/// This function is the inverse of [`from_u32`](Self::from_u32) (see there for more information).
 	#[inline(always)]
@@ -73,22 +74,21 @@ impl Css {
 		u32::from_be_bytes(data)
 	}
 
-	/// Converts a CSS colour to [`Rgba<u8>`].
+	/// Converts an HTML colour to [`SRgba<u8>`].
+	///
+	/// This conversion is always lossless and zero-cost.
 	#[inline(always)]
 	#[must_use]
-	pub const fn to_rgba(self) -> Rgba<u8> {
+	pub const fn to_s_rgba(self) -> SRgba<u8> {
 		self.0
 	}
 }
 
-impl Colour for Css {
-	#[cfg(feature = "wgpu")]
-	fn to_wgpu_color_lossy(&self) -> wgpu_types::Color {
-		self.to_rgba().to_wgpu_color_lossy()
-	}
-}
+impl Colour for Html { }
 
-impl Display for Css {
+impl DefinedGamut for Html { }
+
+impl Display for Html {
 	#[inline]
 	fn fmt(&self, f: &mut Formatter) -> fmt::Result {
 		let value = self.to_u32();
@@ -97,30 +97,38 @@ impl Display for Css {
 	}
 }
 
-impl From<(u8, u8, u8, u8)> for Css {
-	#[inline(always)]
-	fn from((red, green, blue, alpha): (u8, u8, u8, u8)) -> Self {
-		Self::new(red, green, blue, alpha)
-	}
-}
-
-impl From<u32> for Css {
+impl From<u32> for Html {
 	#[inline(always)]
 	fn from(value: u32) -> Self {
 		Self::from_u32(value)
 	}
 }
 
-impl From<Rgba<u8>> for Css {
+impl From<SRgba<u8>> for Html {
 	#[inline(always)]
-	fn from(value: Rgba<u8>) -> Self {
+	fn from(value: SRgba<u8>) -> Self {
 		Self(value)
 	}
 }
 
-impl From<Css> for u32 {
+impl From<SRgb<u8>> for Html {
 	#[inline(always)]
-	fn from(value: Css) -> Self {
+	fn from(value: SRgb<u8>) -> Self {
+		let colour = value.with_alpha(0xFF);
+		Self(colour)
+	}
+}
+
+impl From<Html> for SRgba<u8> {
+	#[inline(always)]
+	fn from(value: Html) -> Self {
+		value.to_s_rgba()
+	}
+}
+
+impl From<Html> for u32 {
+	#[inline(always)]
+	fn from(value: Html) -> Self {
 		value.to_u32()
 	}
 }